Before, we were completing the student flow completely before each scenario. I've tweaked this now, so we create a submittable claim, and then visit the check-your-answers page. I've also cut down on the amount of nested scenarios, which speeds things up further.

This cuts down the amount of time the spec takes to run from ~20 seconds to ~5 seconds, and takes our total test suite run time down from ~1 minute 20s seconds, to ~57 seconds (so under 1 minute! 🙌 )