@pezholio 2019-11-07
![](https://github.trello.services/images/mini-trello-icon.png) [Speed up changing answers feature spec](https://trello.com/c/GWTQ2f3U/974-speed-up-changing-answers-feature-spec)