@pezholio 2019-11-07
@mec Yeah, good call - it seems kinda irelevant given that there's only two choices now